requirements.txt
streamlit
openai
pydantic
azure-cognitiveservices-speech